Refactor Nuvei ACH gateway, replace create_order with create_payment
Modified the existing create_order function in Nuvei ACH gateway to create_payment. This change was made because the create_payment method provides a redirect URL unlike the create_order method. Some unnecessary code was also cleaned up including payment option addition and a method to get transaction id. This streamlines the module and preps it for future functionality for ACH payments.